Prevent arrow heads from vanishing
#912
