+++ waltz-ng/client/logical-flow/components/source-and-target-graph/source-and-target-graph.scss
@@ -1,10 +1,5 @@
 .waltz-source-and-target-graph {
 
-    #wsat-arrowhead {
-        stroke: #bbb;
-        fill: #aaa;
-    }
-
     .wsat-label {
         font-size: 9pt;
     }
