+++ waltz-ng/client/svg-diagram/component/common-svg-defs.html
@@ -1,0 +1,3 @@
+<svg width="0" height="0">
+
+</svg>
