+++ waltz-ng/client/svg-diagram/index.js
@@ -1,7 +1,9 @@
 
 export default (module) => {
-    module.directive('waltzSvgDiagram', require('./directives/svg-diagram'));
-    module.directive('waltzSvgDiagrams', require('./directives/svg-diagrams'));
-    module.service('SvgDiagramStore', require('./services/svg-diagram-store'));
+    module
+        .service('SvgDiagramStore', require('./services/svg-diagram-store'))
+        .directive('waltzSvgDiagram', require('./directives/svg-diagram'))
+        .directive('waltzSvgDiagrams', require('./directives/svg-diagrams'))
+        .component('waltzCommonSvgDefs', require('./component/common-svg-defs'));
 
 };
