+++ waltz-ng/index.template.html
@@ -17,6 +17,8 @@
 
 <body ng-app="waltz-app" ng-strict-di>
 
+    <waltz-common-svg-defs></waltz-common-svg-defs>
+
     <div class="container-fluid">
 
         <!-- Header -->
